--- src/About.fxml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Hyperlink?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.StackPane?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-
-<StackPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/15.0.1" xmlns:fx="http://javafx.com/fxml/1">
-   <children>
-      <ImageView fitHeight="720.0" fitWidth="1280.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@resources/loginBackground.png" />
-         </image>
-      </ImageView>
-      <StackPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="450.0" prefWidth="1080.0" style="-fx-background-color: WHITE;" translateY="80.0">
-         <children>
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Gridflow was developed by 4 Cal Poly Engineering students in their capstone project class." textOrigin="CENTER" translateX="-50.0" translateY="-200.0">
-               <font>
-                  <Font size="24.0" />
-               </font>
-            </Text>
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Hovering over any build menu icons will tell the user what they do." translateX="-175.0" translateY="-160.0">
-               <font>
-                  <Font size="24.0" />
-               </font>
-            </Text>
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="When the build menu is open and active, the following keys are bound:&#10;&#9;A: Pick Association Tool&#10;&#9;S: Pick Selection Tool&#10;&#9;E: Toggle Default Component State&#10;&#9;R: Rotate Component&#10;&#9;ESC: Switch to Interact" translateX="-150.0" translateY="-35.0">
-               <font>
-                  <Font size="24.0" />
-               </font>
-            </Text>
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="For more information regarding this application, please visit our Github and Wiki featured below." translateX="-15.0" translateY="90.0">
-               <font>
-                  <Font size="24.0" />
-               </font>
-            </Text>
-            <Hyperlink style="-fx-border-color: transparent;" text="https://github.com/hrinn/gridflow" translateX="-345.0" translateY="135.0">
-               <font>
-                  <Font size="24.0" />
-               </font>
-            </Hyperlink>
-            <Hyperlink style="-fx-border-color: transparent;" text="https://github.com/hrinn/gridflow/wiki" translateX="-320.0" translateY="180.0">
-               <font>
-                  <Font size="24.0" />
-               </font>
-            </Hyperlink>
-         </children>
-      </StackPane>
-   </children>
-</StackPane>
