--- src/construction/BuildMenuData.java
@@ -1,6 +1,0 @@
-package construction;
-
-public class BuildMenuData {
-    public ToolType toolType = ToolType.INTERACT;
-    public ComponentType componentType;
-}
