--- src/construction/PropertiesMenuView.fxml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<?import javafx.geometry.Insets?>
-
-<BorderPane
-    xmlns="http://javafx.com/javafx"
-    xmlns:fx="http://javafx.com/fxml"
-    fx:controller="construction.PropertiesMenuViewController"
-    prefHeight="600.0" prefWidth="200.0"
-    fx:id="PropertiesWindow" maxHeight="-Infinity" maxWidth="Infinity" minHeight="-Infinity" minWidth="-Infinity" style="-fx-border-color: BLACK; -fx-background-color: RED;">
-    <top>
-    <Label fx:id="Properties" alignment="CENTER" contentDisplay="CENTER" text="Properties" style="-fx-font-size: 15 vw;" BorderPane.alignment="CENTER">
-        <padding>
-            <Insets bottom="5.0" />
-        </padding></Label>
-    </top>
-    <VBox style="-fx-background-color: LIGHTGRAY;">
-
-    </VBox>
-</BorderPane>
