+++ src/construction/buildMenu/BuildMenuData.java
@@ -1,0 +1,9 @@
+package construction.buildMenu;
+
+import construction.ComponentType;
+import construction.ToolType;
+
+public class BuildMenuData {
+    public ToolType toolType = ToolType.INTERACT;
+    public ComponentType componentType;
+}
