+++ src/construction/properties/PropertiesMenuView.fxml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<?import javafx.geometry.Insets?>
+
+<AnchorPane
+    xmlns="http://javafx.com/javafx"
+    xmlns:fx="http://javafx.com/fxml"
+    fx:controller="construction.properties.PropertiesMenuViewController"
+    prefHeight="400.0" prefWidth="200.0"
+    fx:id="PropertiesWindow" maxHeight="-Infinity" maxWidth="Infinity" minHeight="-Infinity" minWidth="-Infinity"
+    styleClass="properties-menu" stylesheets="@propertiesMenuStyle.css">
+    <Label fx:id="Properties" alignment="CENTER" contentDisplay="CENTER" text="Properties" style="-fx-font-size: 18 vw;"
+           AnchorPane.topAnchor="6" AnchorPane.leftAnchor="10"/>
+    <VBox style="-fx-background-color: LIGHTGRAY;">
+
+    </VBox>
+</AnchorPane>
