+++ src/construction/selector/SelectionManagerController.java
@@ -3,7 +3,7 @@ package construction.selector;
 import application.events.GridChangedEvent;
 import application.events.GridFlowEventManager;
 import application.events.SaveStateEvent;
-import construction.BuildMenuData;
+import construction.buildMenu.BuildMenuData;
 import construction.ToolType;
 import construction.canvas.GridCanvasFacade;
 import construction.history.GridMemento;
@@ -11,11 +11,7 @@ import domain.Grid;
 import javafx.collections.ListChangeListener;
 import javafx.event.EventHandler;
 import javafx.scene.Node;
-import javafx.scene.input.KeyCode;
-import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
-
-import java.util.List;
 
 public class SelectionManagerController {
